+++ Code/ID/EffectID.cs
@@ -30,6 +30,7 @@ public static class EffectID
     public const string FlingUpwards = "fling_upwards";
     public const string FartSound = "fart_sound";
     public const string Hiccup = "hiccup";
+    public const string AttackSpeed = "attack_speed";
 
     // --- Buff effects (positive)
     public const string BuffSurvivability = "buff_survivability";